refactor:delete returning value of matchDao use playerservice
